[WIP] Enforce hidden API blacklist
The following patch enforces the hidden API blacklist, preventing access
to boot class path methods/fields marked by the `hiddenapi` tool using
static linking, reflection and JNI.

The mirror::Class::CanAccess function returns `false` when a method
outside boot class path attempt to access a blacklisted member inside
boot class path. As a result, methods which statically link against
hidden members will soft fail bytecode verification and throw
`IllegalAccessError` at run time.

Since the JNI interface has no way of signalling access restrictions
to callers, `JNIEnv::GetMethodID` and `JNIEnv::GetFieldID` now return
`NULL` if the resolved member is blacklisted, hiding it from the caller.
Whether the caller is in boot class path or not is determined using
stack inspection.

Reflection is treated the same as JNI for consistency and blacklisted
members are omitted from reflective queries like `Class::getMethods`.
Looking up a hidden member directly throws `NoSuchFieldException` or
`NoSuchMethodException`. Stack inspection is again used to determine
the class loader membership of the caller.

Test: m test-art-host
Bug: 64382372

+++ runtime/verifier/verifier_deps.h
@@ -126,45 +126,45 @@ class VerifierDeps {
   }
 
  private:
-  static constexpr uint16_t kUnresolvedMarker = static_cast<uint16_t>(-1);
+  static constexpr uint32_t kUnresolvedMarker = static_cast<uint32_t>(-1);
 
-  using ClassResolutionBase = std::tuple<dex::TypeIndex, uint16_t>;
+  using ClassResolutionBase = std::tuple<dex::TypeIndex, uint32_t>;
   struct ClassResolution : public ClassResolutionBase {
     ClassResolution() = default;
     ClassResolution(const ClassResolution&) = default;
-    ClassResolution(dex::TypeIndex type_idx, uint16_t access_flags)
+    ClassResolution(dex::TypeIndex type_idx, uint32_t access_flags)
         : ClassResolutionBase(type_idx, access_flags) {}
 
     bool IsResolved() const { return GetAccessFlags() != kUnresolvedMarker; }
     dex::TypeIndex GetDexTypeIndex() const { return std::get<0>(*this); }
-    uint16_t GetAccessFlags() const { return std::get<1>(*this); }
+    uint32_t GetAccessFlags() const { return std::get<1>(*this); }
   };
 
-  using FieldResolutionBase = std::tuple<uint32_t, uint16_t, dex::StringIndex>;
+  using FieldResolutionBase = std::tuple<uint32_t, uint32_t, dex::StringIndex>;
   struct FieldResolution : public FieldResolutionBase {
     FieldResolution() = default;
     FieldResolution(const FieldResolution&) = default;
-    FieldResolution(uint32_t field_idx, uint16_t access_flags, dex::StringIndex declaring_class_idx)
+    FieldResolution(uint32_t field_idx, uint32_t access_flags, dex::StringIndex declaring_class_idx)
         : FieldResolutionBase(field_idx, access_flags, declaring_class_idx) {}
 
     bool IsResolved() const { return GetAccessFlags() != kUnresolvedMarker; }
     uint32_t GetDexFieldIndex() const { return std::get<0>(*this); }
-    uint16_t GetAccessFlags() const { return std::get<1>(*this); }
+    uint32_t GetAccessFlags() const { return std::get<1>(*this); }
     dex::StringIndex GetDeclaringClassIndex() const { return std::get<2>(*this); }
   };
 
-  using MethodResolutionBase = std::tuple<uint32_t, uint16_t, dex::StringIndex>;
+  using MethodResolutionBase = std::tuple<uint32_t, uint32_t, dex::StringIndex>;
   struct MethodResolution : public MethodResolutionBase {
     MethodResolution() = default;
     MethodResolution(const MethodResolution&) = default;
     MethodResolution(uint32_t method_idx,
-                     uint16_t access_flags,
+                     uint32_t access_flags,
                      dex::StringIndex declaring_class_idx)
         : MethodResolutionBase(method_idx, access_flags, declaring_class_idx) {}
 
     bool IsResolved() const { return GetAccessFlags() != kUnresolvedMarker; }
     uint32_t GetDexMethodIndex() const { return std::get<0>(*this); }
-    uint16_t GetAccessFlags() const { return std::get<1>(*this); }
+    uint32_t GetAccessFlags() const { return std::get<1>(*this); }
     dex::StringIndex GetDeclaringClassIndex() const { return std::get<2>(*this); }
   };
 
@@ -235,7 +235,7 @@ class VerifierDeps {
   // Returns the bytecode access flags of `element` (bottom 16 bits), or
   // `kUnresolvedMarker` if `element` is null.
   template <typename T>
-  static uint16_t GetAccessFlags(T* element)
+  static uint32_t GetAccessFlags(T* element)
       REQUIRES_SHARED(Locks::mutator_lock_);
 
   // Returns a string ID of the descriptor of the declaring class of `element`,
